+++ src/PresentationalComponents/CreateImageWizard/WizardStepReview.scss
@@ -1,0 +1,17 @@
+.error {
+    color: var(--pf-global--danger-color--100);
+  }
+// Increasing margins for h3 for better spacing and readability
+.pf-c-content h3 {
+    margin-top: var(--pf-global--spacer--xl);
+    margin-bottom: var(--pf-global--spacer--md);
+}
+// Decreasing gap between dl items for better spacing and readability
+// Also setting a first column width to 25% instead of auto, 
+// to guarantee the same width for each section.
+@media screen and (min-width: 576px) {
+    .pf-c-content dl {
+        grid-template: 1fr / 25% 1fr;
+        grid-gap: var(--pf-global--spacer--sm);
+    }
+}
